// Banking-FrontEnd/src/app/model/customer.model.ts
export interface Customer {
  customerId: number;
  firstName: string;
  lastName: string;
  emailId: string;
  password: string;
  address: string;
  city: string;
  state: string;
  zipcode: number;
  country: string;
  birthDate: number;
  agree: boolean;
  role: any;
  createdDate:Date;
  status:string;
}